little_arrodes/requirements.txt
discord.py>=2.3.2
google-genai>=0.2.0
python-dotenv>=1.0.1
requests
faiss-cpu
rank-bm25
numpy